chore: switch MUI to use CssVarsProvider and extendTheme
This is the preferred route to take when working with SSR.

CssVarsProvider replaces ThemeProvider

extendTheme replaces createTheme.

NOTE: extendTheme is not a drop in replacement for createTheme.

Docs are currently quite poor for extendTheme, so it will be a suck it and see affair.

The main benefit doing it this way is that we can avoid flickering between dark & light mode on first render.

See: https://mui.com/material-ui/experimental-api/css-theme-variables/overview/#advantages